+++ src/Concordium/Client/Utils.hs
@@ -7,7 +7,11 @@ import Concordium.Crypto.ByteStringHelpers
 import qualified Concordium.ID.Types as IDTypes
 import Data.String.Interpolate (i)
 import qualified Data.Text as Text
+import Data.Word (Word64)
 import Text.Read
+import Data.String.Interpolate (iii)
+import qualified Data.Char as Char
+import Data.Maybe (mapMaybe)
 
 -- | In the 'Left' case of an 'Either', transform the error using the given function and
 -- "rethrow" it in the current 'MonadError'.
@@ -110,3 +114,76 @@ credentialIndexFromStringInform s =
            | otherwise -> Left "Credential index must be non-negative."
     Nothing -> Left "Credential index must be an integer between 0 and 255 (inclusive)."
 
+-- Time Units and durations
+
+secInMs :: Word64
+secInMs = 1000
+
+minInMs :: Word64
+minInMs = 60 * secInMs
+
+hrInMs :: Word64
+hrInMs = 60 * minInMs
+
+dayInMs :: Word64
+dayInMs = 24 * hrInMs
+
+-- |Convert a duration in milliseconds into text with a list of duration measures separated by a space.
+-- A measure is a non-negative integer followed by the unit (with no whitespace in between).
+-- The support units are: days (d), hours (h), minutes (m), seconds (s), milliseconds (ms).
+-- Measures that are 0 are omitted from the output (see example, where 'd' and 'ms' are omitted).
+-- Example: 5022000 -> "1h 23m 42s".
+durationToText :: Word64 -> Text.Text
+durationToText t = Text.intercalate " " . mapMaybe showTimeUnit $
+                  [(d, "d"), (h, "h"), (m, "m"), (s, "s"), (ms, "ms")]
+  where
+    (d, rem0) = quotRem t dayInMs
+    (h, rem1) = quotRem rem0 hrInMs
+    (m, rem2) = quotRem rem1 minInMs
+    (s, ms)   = quotRem rem2 secInMs
+
+    -- Show the time unit if it is non-zero, otherwise return Nothing.
+    showTimeUnit :: (Word64, Text.Text) -> Maybe Text.Text
+    showTimeUnit (value, unit) =
+        if value == 0
+        then Nothing
+        else Just [i|#{value}#{unit}|]
+        
+-- |Parse a string containing a list of duration measures separated by
+-- spaces. A measure is a non-negative integer followed by a unit (no whitespace is allowed in between).
+-- Every measure is accumulated into a duration. The string is allowed to contain
+-- any number of measures with the same unit in no particular order.
+-- The support units are: days (d), hours (h), minutes (m), seconds (s), milliseconds (ms).
+-- Example: "1d 2h 3m 2d 1h" -> Right 270180000
+textToDuration :: Text.Text -> Either String Word64
+textToDuration t = mapM measureToMs measures >>= Right . sum
+  where measures :: [String]
+        measures = map Text.unpack . Text.split (== ' ') $ t
+
+        measureToMs :: String -> Either String Word64
+        measureToMs m = do
+          let (digits, unit) = span Char.isDigit m
+
+          value <- word64FromString digits
+          unit' <- case unit of
+            "ms" -> Right 1
+            "s"  -> Right secInMs
+            "m"  -> Right minInMs
+            "h"  -> Right hrInMs
+            "d"  -> Right dayInMs
+            _    -> Left invalidMeasureErrorMsg
+          Right $ value * unit'
+
+          where
+            invalidMeasureErrorMsg = [iii|"Invalid measure of time '#{m}'.
+                                          Should be a non-negative integer followed by a unit (d, h, m, s, ms)."|]
+
+            -- Reading negative numbers directly to Word64 silently underflows, so this approach is necessary.
+            word64FromString :: String -> Either String Word64
+            word64FromString s = case readMaybe s :: Maybe Integer of
+                                  Nothing -> Left invalidMeasureErrorMsg
+                                  Just x -> if x >= word64MinBound && x <= word64MaxBound
+                                              then Right . fromIntegral $ x
+                                              else Left invalidMeasureErrorMsg
+              where word64MinBound = 0
+                    word64MaxBound = fromIntegral (maxBound :: Word64)